Move AccountSetupFinishedLauncher to feature launcher
