+++ app/k9mail/src/main/java/com/fsck/k9/Dependencies.kt
@@ -8,6 +8,7 @@ import com.fsck.k9.backends.backendsModule
 import com.fsck.k9.controller.ControllerExtension
 import com.fsck.k9.crypto.EncryptionExtractor
 import com.fsck.k9.crypto.openpgp.OpenPgpEncryptionExtractor
+import com.fsck.k9.feature.featureModule
 import com.fsck.k9.notification.notificationModule
 import com.fsck.k9.preferences.K9StoragePersister
 import com.fsck.k9.preferences.StoragePersister
@@ -44,4 +45,5 @@ val appModules = listOf(
     backendsModule,
     storageModule,
     newAccountModule,
+    featureModule,
 )
