+++ app/k9mail/src/main/java/com/fsck/k9/feature/AccountSetupFinishedLauncher.kt
@@ -1,14 +1,14 @@
-package com.fsck.k9.account
+package com.fsck.k9.feature
 
 import android.content.Context
 import android.content.Intent
-import app.k9mail.feature.account.setup.AccountSetupExternalContract
+import app.k9mail.feature.launcher.FeatureLauncherExternalContract
 import com.fsck.k9.activity.MessageList
 
 class AccountSetupFinishedLauncher(
     private val context: Context,
-) : AccountSetupExternalContract.AccountSetupFinishedLauncher {
-    override suspend fun launch(accountUuid: String) {
+) : FeatureLauncherExternalContract.AccountSetupFinishedLauncher {
+    override fun launch(accountUuid: String) {
         val intent = Intent(context, MessageList::class.java).apply {
             addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
             addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK)
