+++ feature/account/setup/src/main/kotlin/app/k9mail/feature/account/setup/AccountSetupExternalContract.kt
@@ -16,8 +16,4 @@ interface AccountSetupExternalContract {
     fun interface AccountOwnerNameProvider {
         suspend fun getOwnerName(): String?
     }
-
-    fun interface AccountSetupFinishedLauncher {
-        suspend fun launch(accountUuid: String)
-    }
 }
