+++ feature/launcher/src/main/kotlin/app/k9mail/feature/launcher/FeatureLauncherExternalContract.kt
@@ -1,0 +1,8 @@
+package app.k9mail.feature.launcher
+
+interface FeatureLauncherExternalContract {
+
+    fun interface AccountSetupFinishedLauncher {
+        fun launch(accountUuid: String)
+    }
+}
